This gets hooks working with cucumber's front-end
Messy, but gives us a sketch to work from.

Notice how Hook implements the same protocol as Test::Step, and
notice how it describes its source, including itself. This allows it to
be printed out when an exception occurs in the hook. It makes me
wonder whether the hook would be better having a source attribute rather
than knowing about the test_case at all.